+++ genomepy/functions.py
@@ -12,7 +12,12 @@ from collections.abc import Iterable
 from pyfaidx import Fasta, Sequence
 from genomepy.provider import ProviderBase
 from genomepy.plugin import get_active_plugins, init_plugins
-from genomepy.utils import generate_gap_bed, generate_fa_sizes, get_localname
+from genomepy.utils import (
+    generate_gap_bed,
+    generate_fa_sizes,
+    get_localname,
+    sanitize_annotation,
+)
 from genomepy.exceptions import GenomeDownloadError
 
 config = norns.config("genomepy", default="cfg/default.yaml")
@@ -158,6 +163,9 @@ def install_genome(
     invert_match=False,
     bgzip=None,
     annotation=False,
+    only_annotation=False,
+    skip_sanitizing=False,
+    threads=1,
     force=False,
     **kwargs,
 ):
@@ -191,26 +199,42 @@ def install_genome(
         If set to True the genome FASTA file will be compressed using bgzip.
         If not specified, the setting from the configuration file will be used.
 
-    annotation : bool , optional
-        If set to True, download gene annotation in BED and GTF format.
+    threads : int, optional
+        Build genome index using multithreading (if supported).
 
     force : bool , optional
         Set to True to overwrite existing files.
 
+    annotation : bool , optional
+        If set to True, download gene annotation in BED and GTF format.
+
+    only_annotation : bool , optional
+        If set to True, only download the annotation files.
+
+    skip_sanitizing : bool , optional
+        If set to True, downloaded annotation files whose sequence names do not match
+        with the (first header fields of) the genome.fa will not be corrected.
+
     kwargs : dict, optional
         Provider specific options.
-        Ensembl:
-
         toplevel : bool , optional
             Ensembl only: Always download the toplevel genome. Ignores potential primary assembly.
 
         version : int, optional
             Ensembl only: Specify release version. Default is latest.
+
+        to_annotation : text , optional
+            URL only: direct link to annotation file.
+            Required if this is not the same directory as the fasta.
     """
     if not genome_dir:
         genome_dir = config.get("genome_dir", None)
     if not genome_dir:
         raise norns.exceptions.ConfigError("Please provide or configure a genome_dir")
+
+    # download annotation if any of the annotation related flags are given
+    if only_annotation or kwargs.get("to_annotation", False):
+        annotation = True
 
     genome_dir = os.path.expanduser(genome_dir)
     localname = get_localname(name, localname)
@@ -220,7 +244,7 @@ def install_genome(
     no_genome_found = not any(
         os.path.exists(fname) for fname in glob_ext_files(out_dir, "fa")
     )
-    if no_genome_found or force:
+    if (no_genome_found or force) and not only_annotation:
         # Download genome from provider
         p = ProviderBase.create(provider)
         p.download_genome(
@@ -234,31 +258,45 @@ def install_genome(
             **kwargs,
         )
 
-    # If annotation is requested, check if annotation already exists, or if downloading is forced
+    # annotation_only cannot use sanitizing if no genome (and sizes) file was made earlier. Warn the user about this.
     no_annotation_found = not any(
         os.path.exists(fname) for fname in glob_ext_files(out_dir, "gtf")
     )
-    if annotation and (no_annotation_found or force):
-        # Download annotation from provider
-        p = ProviderBase.create(provider)
-        p.download_annotation(name, genome_dir, localname=localname, **kwargs)
+    no_genome_found = not any(
+        os.path.exists(fname) for fname in glob_ext_files(out_dir, "fa")
+    )
+    if only_annotation and no_genome_found:
+        assert skip_sanitizing, (
+            "a genome file is required to sanitize your annotation (or check if it's required). "
+            "Use the skip sanitizing flag (-s) if you wish to skip this step."
+        )
 
     # generates a Fasta object and the index file
-    g = Genome(localname, genome_dir=genome_dir)
-
-    # Run all active plugins
-    for plugin in get_active_plugins():
-        plugin.after_genome_download(g, force)
-
-    # Generate gap file if not found or if generation is forced
-    gap_file = os.path.join(out_dir, localname + ".gaps.bed")
-    if not os.path.exists(gap_file) or force:
-        generate_gap_bed(glob_ext_files(out_dir, "fa")[0], gap_file)
+    if not no_genome_found:
+        g = Genome(localname, genome_dir=genome_dir)
 
     # Generate sizes file if not found or if generation is forced
     sizes_file = os.path.join(out_dir, localname + ".fa.sizes")
-    if not os.path.exists(sizes_file) or force:
+    if (not os.path.exists(sizes_file) or force) and not only_annotation:
         generate_fa_sizes(glob_ext_files(out_dir, "fa")[0], sizes_file)
+
+    # Generate gap file if not found or if generation is forced
+    gap_file = os.path.join(out_dir, localname + ".gaps.bed")
+    if (not os.path.exists(gap_file) or force) and not only_annotation:
+        generate_gap_bed(glob_ext_files(out_dir, "fa")[0], gap_file)
+
+    # If annotation is requested, check if annotation already exists, or if downloading is forced
+    if (no_annotation_found or force) and annotation:
+        # Download annotation from provider
+        p = ProviderBase.create(provider)
+        p.download_annotation(name, genome_dir, localname=localname, **kwargs)
+        if not skip_sanitizing:
+            sanitize_annotation(g)
+
+    # Run all active plugins
+    for plugin in get_active_plugins():
+        plugin.after_genome_download(g, threads, force)
+
     generate_env()