+++ CHANGELOG.md
@@ -8,12 +8,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Blacklists are automatically unzipped.
+- Downloading of annotation file (BED/GTF/GFF3) from URL
+- Automatic search for annotation file (GTF/GFF3) in genome directory when downloading from URL
+- Annotation sanitizing (and skip sanitizing flag)
 
 ### Changed
 - sizes no longer a plugin, but always gets executed
+- `genomepy install --help` menu now contains paragraphs
 
 ### Fixes
 - Fix minor issue with hg19 wrong blacklist url
+- tests now use double digits to preserve testing order
+- ensembl downloads over http instead of https (as release 99 does no longer have https)
 
 ## [0.7.1] - 2019-11-20